# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KIN API",
    "version": "1.0.0",
    "description": "KIN AI Companion Platform — We Build You A Friend\n\nThe KIN API powers Mission Control (web dashboard), Telegram/Discord/WhatsApp bots, local KIN instances, and third-party integrations. It provides:\n\n- **Auth** — Telegram Login Widget + JWT-based sessions\n- **Chat** — Multi-companion AI chat with a two-brain supervisor (Local Ollama → Groq → Anthropic/OpenAI fallback)\n- **Skills Marketplace** — Browse, install, and custom-submit companion skill extensions\n- **Companion Skills** — NFT-portable XP/leveling system for companion skill trees\n- **Heartbeat & Recovery** — Local KIN health monitoring and snapshot-based recovery\n- **Billing** — Stripe subscription checkout, portal, and webhook handling\n- **Support** — AI-powered support chat with human escalation, FAQ, tickets, and feature requests\n\nAll protected routes require a JWT bearer token obtained via `POST /auth/telegram` or `POST /auth/dev-login` (development only).",
    "contact": {
      "name": "KIN Support",
      "url": "https://www.meetyourkin.com/support",
      "email": "support@meetyourkin.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/kr8tiv-ai/kr8tiv-runtime-truth-contracts/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    },
    {
      "url": "https://api.meetyourkin.com",
...
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token. Obtain via `POST /auth/telegram` or `POST /auth/dev-login`. Token expires after 2 hours. Algorithm: HS256."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "oneOf": [
              { "type": "string" },
              {
                "type": "object",
                "properties": {
                  "message": { "type": "string" },
                  "statusCode": { "type": "integer" }
                }
              }
            ]
          }
        },
        "example": { "error": "Unauthorized" }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" }
        },
        "example": { "error": "Bad Request", "message": "Message is required" }
      },
...
      "CompanionId": {
        "type": "string",
        "enum": ["cipher", "mischief", "vortex", "forge", "aether", "catalyst"],
        "description": "Companion identifier. Each companion has a distinct personality and specialisation.",
        "example": "cipher"
      },
...
      "CompanionSkillEntry": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "companionId": { "$ref": "#/components/schemas/CompanionId" },
          "skillId": { "type": "string" },
          "skillName": { "type": "string" },
          "skillDisplayName": { "type": "string" },
          "skillLevel": { "type": "integer", "minimum": 1, "maximum": 99 },
          "xp": { "type": "integer" },
          "xpToNextLevel": { "type": "integer" },
          "isPortable": {
            "type": "boolean",
            "description": "Portable skills travel with the NFT when the companion is transferred."
          },
          "usageCount": { "type": "integer" },
          "accruedAt": { "type": "string", "format": "date-time" },
          "lastUsedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT token.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Unauthorized" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request body or parameters.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ValidationError" }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limit exceeded. Default: 100 requests/minute per IP in production.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": { "type": "integer", "example": 429 },
                "error": { "type": "string", "example": "Too Many Requests" },
                "message": { "type": "string" }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [],
...
  "paths": {
...
    "/companion-skills/{companionId}": {
      "get": {
        "tags": ["Companions"],
        "operationId": "companionSkillsGet",
        "summary": "Get companion skills",
        "description": "Returns all accrued skills for a specific companion, ordered by level descending. Each skill entry includes XP, level, and portability status.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "companionId",
            "in": "path",
            "required": true,
            "schema": { "$ref": "#/components/schemas/CompanionId" }
          }
        ],
        "responses": {
          "200": {
            "description": "Companion's skill entries.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CompanionSkillEntry" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/companion-skills/{companionId}/accrue": {
      "post": {
        "tags": ["Companions"],
        "operationId": "companionSkillsAccrue",
        "summary": "Award XP to companion",
        "description": "Awards experience points (XP) to a companion's skill. Automatically handles level-up when XP crosses the threshold. XP values follow the curve: [0, 100, 300, 600, 1000...] per level.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "companionId",
            "in": "path",
            "required": true,
            "schema": { "$ref": "#/components/schemas/CompanionId" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["skillId"],
                "properties": {
                  "skillId": {
                    "type": "string",
                    "description": "Skill ID to award XP to.",
                    "example": "skill-abc123"
                  },
                  "xpGained": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100,
                    "default": 10,
                    "description": "XP to award. Capped at 100 per call."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated skill state.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "skillLevel": { "type": "integer" },
                    "xp": { "type": "integer" },
                    "xpToNextLevel": { "type": "integer" },
                    "leveledUp": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/companion-skills/snapshot": {
      "post": {
        "tags": ["Companions"],
        "operationId": "companionSkillsSnapshot",
        "summary": "Create skill snapshot",
        "description": "Creates a SHA-256-hashed snapshot of a companion's portable skills. Used before NFT transfers and for IPFS/on-chain skill state attestation. Private memories and conversation history are never included.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["companionId"],
                "properties": {
                  "companionId": { "$ref": "#/components/schemas/CompanionId" },
                  "snapshotType": {
                    "type": "string",
                    "enum": ["skill_state", "personality", "full", "transfer"],
                    "default": "skill_state"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Snapshot created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "contentHash": { "type": "string", "description": "SHA-256 hex digest of the skill payload." },
                    "skillCount": { "type": "integer" },
                    "totalLevels": { "type": "integer" },
                    "snapshotType": { "type": "string" },
                    "ipfsCid": { "type": "string", "nullable": true },
                    "isOnChain": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/companion-skills/snapshots": {
      "get": {
        "tags": ["Companions"],
        "operationId": "companionSkillsSnapshots",
        "summary": "List snapshots",
        "description": "Lists skill snapshots for the authenticated user, optionally filtered by companion. Returns up to 50 most recent snapshots.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "companionId",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/CompanionId" },
            "description": "Filter snapshots to a specific companion."
          }
        ],
        "responses": {
          "200": {
            "description": "List of snapshots.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "companionId": { "type": "string" },
                      "snapshotType": { "type": "string" },
                      "contentHash": { "type": "string" },
                      "ipfsCid": { "type": "string", "nullable": true },
                      "solanaTxSig": { "type": "string", "nullable": true },
                      "isOnChain": { "type": "boolean" },
                      "createdAt": { "type": "string", "format": "date-time" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/companion-skills/transfer": {
      "post": {
        "tags": ["Companions"],
        "operationId": "companionSkillsTransfer",
        "summary": "Prepare NFT transfer",
        "description": "Prepares a companion's portable skills for NFT transfer. Creates a transfer snapshot, logs the transfer, and—if `toUserId` is provided—replicates the portable skills to the new owner. Private memories and conversation history remain with the original owner and are never transferred.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["companionId", "nftMintAddress"],
                "properties": {
                  "companionId": { "$ref": "#/components/schemas/CompanionId" },
                  "nftMintAddress": {
                    "type": "string",
                    "description": "Solana mint address of the companion NFT.",
                    "example": "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
                  },
                  "toUserId": {
                    "type": "string",
                    "description": "KIN user ID of the recipient. If provided, skills are replicated to the new owner.",
                    "example": "user-xyz789"
                  },
                  "transferTxSig": {
                    "type": "string",
                    "description": "Solana transaction signature for the NFT transfer (for audit log).",
                    "example": "5VfYjH..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transfer prepared.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transferId": { "type": "string" },
                    "snapshotId": { "type": "string" },
                    "contentHash": { "type": "string" },
                    "skillsTransferred": { "type": "integer" },
                    "totalLevels": { "type": "integer" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": {
            "description": "Forbidden — caller does not own this companion NFT.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/companion-skills/transfer-log": {
      "get": {
        "tags": ["Companions"],
        "operationId": "companionSkillsTransferLog",
        "summary": "Transfer history",
        "description": "Returns all NFT transfer events where the authenticated user is either sender or recipient. Limited to 50 most recent records.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Transfer log entries.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "nftMintAddress": { "type": "string" },
                      "companionId": { "type": "string" },
                      "fromUserId": { "type": "string" },
                      "toUserId": { "type": "string", "nullable": true },
                      "skillsTransferred": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "skillId": { "type": "string" },
                            "level": { "type": "integer" },
                            "xp": { "type": "integer" },
                            "name": { "type": "string" }
                          }
                        }
                      },
                      "transferTxSig": { "type": "string", "nullable": true },
                      "createdAt": { "type": "string", "format": "date-time" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
...
  }
}